// src/store/main/system/types.ts
/*
 * @Description: 待编辑
 * @Author: SiFeng Zhai
 * @Date: 2022-05-26 16:57:45
 * @LastEditors: SiFeng Zhai
 * @LastEditTime: 2022-08-09 22:13:12
 */
export interface ISystemState {
  usersList: any[]
  usersCount: number
  roleList: any[]
  roleCount: number
  goodsList: any[]
  goodsCount: number
  menuList: any[]
  menuCount: number
}